Validation: created validation for categories and products.

+++ app/Http/Controllers/Admin/Categories/CategoriesController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Admin\Categories;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
-use Illuminate\Http\Request;
+use App\Http\Requests\CategoriesRequest;
 use Illuminate\Support\Facades\Storage;
 
 class CategoriesController extends Controller
@@ -37,11 +37,14 @@ class CategoriesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoriesRequest $request)
     {
-        $path = $request->file('image')->store('public/categories');
         $params = $request->all();
-        $params['image'] = $path;
+        if($request->has('image')){
+            $path = $request->file('image')->store('public/categories');
+            $params['image'] = $path;
+        }
+       
         $category = Category::create($params);
 
         return redirect()->route('categories.index');
@@ -77,12 +80,15 @@ class CategoriesController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoriesRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('public/categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if($request->has('image')){
+            Storage::delete($category->image);
+            $path = $request->file('image')->store('public/categories');
+            $params['image'] = $path;
+        }
 
         $category->update($params);
